Add vendor hook to handle_control_message
This is needed for Nvidia Shield devices to handle a
'restart consolemode' request from a blobbed stock app.

The vendor function is expected to return non-0 if it did
not handle the message and 0 if it did (or wants it ignored).

Forward port of 119414.

Change-Id: Ia8e4ba0bbf561f29f72862cd986f1660d7b501da

init: Add missing include

Change-Id: I3398cf9bae80a5b2ab182fe732b78b1caf1eba15

init: vendor_init: Remove duplicated {

Change-Id: Ib098a0b9e5744b07854a23206de0f36fe74e0ed7